refactor(common/utils): move the deferred utility to its own file because it will be needed by itself
