+++ src/common/utils/deferred.js
@@ -1,0 +1,14 @@
+export const deferred = () => {
+  let resolve;
+  let reject;
+  const promise = new Promise((presolve, preject) => {
+    resolve = presolve;
+    reject = preject;
+  });
+
+  return {
+    promise,
+    resolve,
+    reject
+  };
+};
